requirements.txt
requests
PyPDF2
python-docx
tiktoken
numpy